frontend/requirements.txt
fastapi
uvicorn
streamlit
requests
google-generativeai
python-multipart
numpy
pydantic
chromadb
pypdf
langchain-core
langchain-text-splitters
langchain